all components under components/ui folder renamed to pascal-case to avoid confusion
