+++ app/src/components/DisplayContent.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import DisplayHeadings from './DisplayHeadings';
-import Divider from './ui/divider';
+import Divider from './ui/CustomDivider';
 import RightSidebarPortal from './RightSidebarPortal';
 import { getContent } from '@/lib';
 import { FiEdit } from '@/data';
