+++ app/src/components/ui/Home/Hero.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import { Button } from '../Button';
+import { Button } from '../CustomButton';
 import { FaGithub } from '@/data';
 
 const Hero = () => {
